pyproject.toml
[tool.poetry]
name = "crosslink"
version = "0.1.0"
description = "CrossLink"
authors = ["Tien Nguyen <vantien.nguyen.fr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
django-environ = "^0.11.2"
django-sslserver = "^0.22"
psycopg2 = "^2.9.9"
django-extensions = "^3.2.3"
Werkzeug = "^3.0.2"
djangorestframework = "^3.15.1"
ipython = "^8.23.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^23.5"
pre-commit = "^3.7.0"
isort = "^5.13.2"
black = "^24.4.0"
factory-boy = "^3.3.0"
pyright = "^1.1.358"
ShopifyAPI = "^12.5.0"
django-cors-headers = "^4.3.1"
celery = "^5.3.6"
redis = "^5.0.3"
gunicorn = "^21.2.0"
autoflake = "^2.3.1"
numpy = "^1.26.4"
pytz = "^2024.1"
boto3 = "^1.34.141"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.pyright]
typeCheckingMode = "off"
reportUnusedImport = true
